+++ src/vnctools/vnctools-kill.sh
@@ -3,8 +3,12 @@ function main() {
     bashargs::add_required_value --display
     bashargs::add_required_value --hostname
     bashargs::add_required_value --username
+    bashargs::add_optional_flag --trace
     bashargs::parse_args $@
 
+    if [[ $(bashargs::get_arg --trace) = true ]]; then
+        set -x
+    fi
     ssh $(bashargs::get_arg --username)@$(bashargs::get_arg --hostname) \
         "set -x
         vncserver -kill :$(bashargs::get_arg --display)